requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
faiss-cpu==1.7.4
numpy>=1.26.2
pandas==2.0.3
PyPDF2==3.0.1
python-docx==0.8.11
sentence-transformers>=2.7.0
transformers>=4.21.0
huggingface_hub>=0.16.0
torch>=1.9.0
plotly==5.17.0
streamlit-option-menu==0.3.6
python-dotenv==1.0.0
scikit-learn>=1.0.0